add account service

+++ tonlibjson-jsonrpc/src/main.rs
@@ -19,7 +19,7 @@ use crate::view::{BlockHeaderView, BlockIdExtView, MasterchainInfoView, ShardsVi
 
 #[derive(Deserialize, Debug)]
 struct LookupBlockParams {
-    workchain: i64,
+    workchain: i32,
     shard: String,
     seqno: Option<i32>,
     lt: Option<i64>,
@@ -34,7 +34,7 @@ struct ShardsParams {
 #[allow(dead_code)]
 #[derive(Deserialize)]
 struct BlockHeaderParams {
-    workchain: i64,
+    workchain: i32,
     shard: String,
     seqno: i32,
     root_hash: Option<String>,
@@ -44,7 +44,7 @@ struct BlockHeaderParams {
 #[allow(dead_code)]
 #[derive(Deserialize)]
 struct BlockTransactionsParams {
-    workchain: i64,
+    workchain: i32,
     shard: String,
     seqno: i32,
     root_hash: Option<String>,
